Extra check before removing temp file

+++ library/Exakat/Data/Methods.php
@@ -42,7 +42,7 @@ class Methods {
     }
 
     public function __destruct() {
-        if ($this->phar_tmp !== null) {
+        if ($this->phar_tmp !== null && file_exists($this->phar_tmp)) {
             unlink($this->phar_tmp);
         }
     }
